Read data from 4x VL53L3CX and 4x VL53L5CX
Configure 8 sensors before FreeRTOS startup.

+++ User/Interfaces/Inc/sensor_interface.h
@@ -13,7 +13,7 @@ typedef enum {
 } Sensor_Status;
 
 typedef struct {
-    uint16_t distances[MAX_SENSOR_ROWS][MAX_SENSOR_COLS];
+    int32_t distances[MAX_SENSOR_ROWS][MAX_SENSOR_COLS];
     uint8_t rows;
     uint8_t cols;
 } Sensor_Data;
